requirements.txt
python-dotenv==1.1.0
Flask==3.1.0
Werkzeug==3.1.3
Flask_Login==0.6.3
Flask_CKEditor==1.0.0
Flask-Gravatar==0.5.0
Flask_WTF==1.2.2
WTForms==3.2.1
SQLAlchemy==2.0.40
gunicorn==23.0.0
psycopg2-binary==2.9.10
email-validator==2.2.0
flask-sqlalchemy==3.1.1
Flask-Bootstrap==3.3.7.1
Bootstrap-Flask==2.4.2
bleach==6.2.0